Fix Shaman Maelstrom Weapon

+++ sim/shaman/fire_totems.go
@@ -42,7 +42,7 @@ func (shaman *Shaman) newSearingTotemSpellConfig(rank int) core.SpellConfig {
 
 	spell := core.SpellConfig{
 		ActionID:    core.ActionID{SpellID: spellId},
-		SpellCode:   int32(SpellCode_SearingTotem),
+		SpellCode:   SpellCode_SearingTotem,
 		SpellSchool: core.SpellSchoolFire,
 		ProcMask:    core.ProcMaskEmpty,
 		Flags:       SpellFlagTotem | core.SpellFlagAPL,
@@ -83,10 +83,10 @@ func (shaman *Shaman) newSearingTotemSpellConfig(rank int) core.SpellConfig {
 		},
 
 		ApplyEffects: func(sim *core.Simulation, _ *core.Unit, spell *core.Spell) {
-			if shaman.ActiveTotems[FireTotem] != nil && shaman.ActiveTotems[FireTotem].SpellCode == int32(SpellCode_MagmaTotem) {
+			if shaman.ActiveTotems[FireTotem] != nil && shaman.ActiveTotems[FireTotem].SpellCode == SpellCode_MagmaTotem {
 				shaman.ActiveTotems[FireTotem].AOEDot().Cancel(sim)
 			}
-			if shaman.ActiveTotems[FireTotem] != nil && shaman.ActiveTotems[FireTotem].SpellCode == int32(SpellCode_FireNovaTotem) {
+			if shaman.ActiveTotems[FireTotem] != nil && shaman.ActiveTotems[FireTotem].SpellCode == SpellCode_FireNovaTotem {
 				shaman.ActiveTotems[FireTotem].AOEDot().Cancel(sim)
 			}
 			spell.Dot(sim.GetTargetUnit(0)).Apply(sim)
@@ -131,7 +131,7 @@ func (shaman *Shaman) newMagmaTotemSpellConfig(rank int) core.SpellConfig {
 
 	spell := core.SpellConfig{
 		ActionID:    core.ActionID{SpellID: spellId},
-		SpellCode:   int32(SpellCode_MagmaTotem),
+		SpellCode:   SpellCode_MagmaTotem,
 		SpellSchool: core.SpellSchoolFire,
 		ProcMask:    core.ProcMaskEmpty,
 		Flags:       SpellFlagTotem | core.SpellFlagAPL,
@@ -172,10 +172,10 @@ func (shaman *Shaman) newMagmaTotemSpellConfig(rank int) core.SpellConfig {
 		},
 
 		ApplyEffects: func(sim *core.Simulation, _ *core.Unit, spell *core.Spell) {
-			if shaman.ActiveTotems[FireTotem] != nil && shaman.ActiveTotems[FireTotem].SpellCode == int32(SpellCode_SearingTotem) {
+			if shaman.ActiveTotems[FireTotem] != nil && shaman.ActiveTotems[FireTotem].SpellCode == SpellCode_SearingTotem {
 				shaman.ActiveTotems[FireTotem].Dot(shaman.CurrentTarget).Cancel(sim)
 			}
-			if shaman.ActiveTotems[FireTotem] != nil && shaman.ActiveTotems[FireTotem].SpellCode == int32(SpellCode_FireNovaTotem) {
+			if shaman.ActiveTotems[FireTotem] != nil && shaman.ActiveTotems[FireTotem].SpellCode == SpellCode_FireNovaTotem {
 				shaman.ActiveTotems[FireTotem].AOEDot().Cancel(sim)
 			}
 			spell.AOEDot().Apply(sim)
@@ -227,7 +227,7 @@ func (shaman *Shaman) newFireNovaTotemSpellConfig(rank int) core.SpellConfig {
 	spell := core.SpellConfig{
 		ActionID:    core.ActionID{SpellID: spellId},
 		SpellSchool: core.SpellSchoolFire,
-		SpellCode:   int32(SpellCode_FireNovaTotem),
+		SpellCode:   SpellCode_FireNovaTotem,
 		ProcMask:    core.ProcMaskEmpty,
 		Flags:       SpellFlagTotem | core.SpellFlagAPL,
 
@@ -271,10 +271,10 @@ func (shaman *Shaman) newFireNovaTotemSpellConfig(rank int) core.SpellConfig {
 		},
 
 		ApplyEffects: func(sim *core.Simulation, _ *core.Unit, spell *core.Spell) {
-			if shaman.ActiveTotems[FireTotem] != nil && shaman.ActiveTotems[FireTotem].SpellCode == int32(SpellCode_SearingTotem) {
+			if shaman.ActiveTotems[FireTotem] != nil && shaman.ActiveTotems[FireTotem].SpellCode == SpellCode_SearingTotem {
 				shaman.ActiveTotems[FireTotem].Dot(shaman.CurrentTarget).Cancel(sim)
 			}
-			if shaman.ActiveTotems[FireTotem] != nil && shaman.ActiveTotems[FireTotem].SpellCode == int32(SpellCode_MagmaTotem) {
+			if shaman.ActiveTotems[FireTotem] != nil && shaman.ActiveTotems[FireTotem].SpellCode == SpellCode_MagmaTotem {
 				shaman.ActiveTotems[FireTotem].AOEDot().Cancel(sim)
 			}
 			spell.AOEDot().Apply(sim)
